requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alibabacloud-credentials==0.3.2
alibabacloud-dysmsapi20170525==2.0.23
alibabacloud-endpoint-util==0.0.3
alibabacloud-gateway-spi==0.0.1
alibabacloud-openapi-util==0.2.1
alibabacloud-tea==0.3.2
alibabacloud-tea-openapi==0.3.7
alibabacloud-tea-util==0.3.8
alibabacloud-tea-xml==0.0.2
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.16.1
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
crcmod==1.7
cryptography==41.0.0
Django==4.2.1
django-mysql==4.9.0
django-redis==5.2.0
django-rest-swagger==2.2.0
djangorestframework==3.14.0
frozenlist==1.3.3
idna==3.4
install==1.3.5
itypes==1.2.0
Jinja2==3.1.2
jmespath==0.10.0
MarkupSafe==2.1.2
multidict==6.0.4
mysqlclient==2.1.1
openapi-codec==1.3.2
oss2==2.18.0
Pillow==9.5.0
pycparser==2.21
pycryptodome==3.18.0
pytz==2023.3
redis==4.5.5
requests==2.31.0
simplejson==3.19.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.6.2
uritemplate==4.1.1
urllib3==2.0.2
yarl==1.9.2
